Move summary() to config.py
Also removes check_script_input() and its usages since we already
check scripts during config parsing.
